criacao do modulo exports e limpeza da view
